📦 NEW: Layers + Create proof and Verify
